+++ src/draw.rs
@@ -4,104 +4,7 @@ use crate::color::{
     Color, ColorExt,
     colors::{BLACK, WHITE},
 };
-
-/// A rectangle defined by position and size
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct Rect {
-    pub pos: Vec2,
-    pub size: Vec2,
-}
-
-impl Rect {
-    pub fn new(x: f32, y: f32, width: f32, height: f32) -> Self {
-        Self {
-            pos: Vec2::new(x, y),
-            size: Vec2::new(width, height),
-        }
-    }
-
-    pub fn from_pos_size(pos: Vec2, size: Vec2) -> Self {
-        Self { pos, size }
-    }
-
-    pub fn min(&self) -> Vec2 {
-        self.pos
-    }
-
-    pub fn max(&self) -> Vec2 {
-        self.pos + self.size
-    }
-
-    pub fn contains(&self, point: Vec2) -> bool {
-        point.x >= self.pos.x
-            && point.y >= self.pos.y
-            && point.x <= self.pos.x + self.size.x
-            && point.y <= self.pos.y + self.size.y
-    }
-
-    pub fn intersect(&self, other: &Rect) -> Option<Rect> {
-        let min = self.min().max(other.min());
-        let max = self.max().min(other.max());
-
-        if min.x < max.x && min.y < max.y {
-            Some(Rect::from_pos_size(min, max - min))
-        } else {
-            None
-        }
-    }
-
-    /// Check if this rectangle is fully contained within another rectangle
-    pub fn is_contained_in(&self, other: &Rect) -> bool {
-        self.pos.x >= other.pos.x
-            && self.pos.y >= other.pos.y
-            && self.pos.x + self.size.x <= other.pos.x + other.size.x
-            && self.pos.y + self.size.y <= other.pos.y + other.size.y
-    }
-
-    /// Check if this rectangle is partially visible within another rectangle
-    pub fn is_partially_visible_in(&self, other: &Rect) -> bool {
-        self.intersect(other).is_some()
-    }
-
-    /// Get the area of this rectangle
-    pub fn area(&self) -> f32 {
-        self.size.x * self.size.y
-    }
-
-    /// Get the visibility ratio when clipped by another rectangle (0.0 to 1.0)
-    pub fn visibility_ratio_in(&self, other: &Rect) -> f32 {
-        if let Some(intersection) = self.intersect(other) {
-            let intersection_area = intersection.area();
-            let total_area = self.area();
-            if total_area > 0.0 {
-                intersection_area / total_area
-            } else {
-                0.0
-            }
-        } else {
-            0.0
-        }
-    }
-
-    /// Expand the rectangle by the given amount in all directions
-    pub fn expand(&self, amount: f32) -> Rect {
-        Rect::from_pos_size(
-            self.pos - Vec2::splat(amount),
-            self.size + Vec2::splat(amount * 2.0),
-        )
-    }
-
-    /// Expand the rectangle by the given vector amount
-    pub fn expand_by(&self, expansion: Vec2) -> Rect {
-        Rect::from_pos_size(self.pos - expansion, self.size + expansion * 2.0)
-    }
-
-    /// Contract the rectangle by the given amount in all directions
-    pub fn contract(&self, amount: f32) -> Rect {
-        let new_size = (self.size - Vec2::splat(amount * 2.0)).max(Vec2::ZERO);
-        Rect::from_pos_size(self.pos + Vec2::splat(amount), new_size)
-    }
-}
+use crate::geometry::Rect;
 
 /// Text styling information
 #[derive(Debug, Clone, PartialEq)]
