+++ src/platform/mac.rs
@@ -1,4 +1,5 @@
 mod menu;
+pub(crate) mod metal_renderer;
 mod window;
 
 pub use menu::create_app_menu;
